requirements.txt
certifi==2023.11.17
charset-normalizer==3.3.2
idna==3.6
loguru==0.7.2
python-dotenv==1.0.0
requests==2.31.0
selectolax==0.3.17
sentry-sdk==1.39.2
urllib3==2.1.0
playwright==1.51.0
selenium==4.31.0
trio==0.29.0
trio-websocket==0.12.2
websocket-client==1.8.0